agents/crew-ai/requirements.txt
crewai[tools]>=0.11.0
python-dotenv>=0.19.0
jsonschema>=4.21.1